pyproject.toml
[tool.poetry]
name = "ncbirequest"
version = "0.1.0"
description = ""
authors = ["Andreas Leitner <email@andreas-leitner.at>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
eel = "^0.16.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
